+++ bitwarden_license/test/Commercial.Core.Test/Billing/TaxServiceTests.cs
@@ -1,4 +1,5 @@
 ﻿using Bit.Core.Billing.Services;
+using Bit.Core.Billing.Tax.Services;
 using Bit.Test.Common.AutoFixture;
 using Bit.Test.Common.AutoFixture.Attributes;
 using Xunit;
