+++ src/Api/Billing/Controllers/OrganizationBillingController.cs
@@ -12,6 +12,7 @@ using Bit.Core.Billing.Payment.Services;
 using Bit.Core.Billing.Pricing;
 using Bit.Core.Billing.Providers.Services;
 using Bit.Core.Billing.Services;
+using Bit.Core.Billing.Tax.Models;
 using Bit.Core.Context;
 using Bit.Core.Repositories;
 using Bit.Core.Services;
