+++ src/Api/Billing/Models/Responses/PaymentMethodResponse.cs
@@ -1,5 +1,6 @@
 ﻿using Bit.Core.Billing.Models;
 using Bit.Core.Billing.Payment.Models;
+using Bit.Core.Billing.Tax.Models;
 
 namespace Bit.Api.Billing.Models.Responses;
