--- src/Core/Billing/Models/Api/Responses/PreviewInvoiceResponseModel.cs
@@ -1,7 +1,0 @@
-﻿namespace Bit.Core.Billing.Models.Api.Responses;
-
-public record PreviewInvoiceResponseModel(
-    decimal EffectiveTaxRate,
-    decimal TaxableBaseAmount,
-    decimal TaxAmount,
-    decimal TotalAmount);
