+++ src/Core/Billing/Organizations/Requests/PreviewOrganizationInvoiceRequestModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using Bit.Core.Billing.Enums;
-using Bit.Core.Billing.Models.Api.Requests;
+using Bit.Core.Billing.Tax.Requests;
 using Bit.Core.Enums;
 
 namespace Bit.Core.Billing.Organizations.Requests;
