+++ src/Core/Billing/Organizations/Services/IOrganizationBillingService.cs
@@ -2,6 +2,7 @@
 using Bit.Core.Billing.Models;
 using Bit.Core.Billing.Organizations.Models;
 using Bit.Core.Billing.Payment.Models;
+using Bit.Core.Billing.Tax.Models;
 
 namespace Bit.Core.Billing.Organizations.Services;
