+++ src/Core/Billing/Organizations/Services/OrganizationBillingService.cs
@@ -8,7 +8,8 @@ using Bit.Core.Billing.Payment.Models;
 using Bit.Core.Billing.Payment.Services;
 using Bit.Core.Billing.Pricing;
 using Bit.Core.Billing.Services;
-using Bit.Core.Billing.Services.Contracts;
+using Bit.Core.Billing.Tax.Models;
+using Bit.Core.Billing.Tax.Services;
 using Bit.Core.Enums;
 using Bit.Core.Exceptions;
 using Bit.Core.Repositories;
