+++ src/Core/Billing/Payment/Models/PaymentMethod.cs
@@ -1,4 +1,5 @@
 ﻿using Bit.Core.Billing.Models;
+using Bit.Core.Billing.Tax.Models;
 
 namespace Bit.Core.Billing.Payment.Models;
