+++ src/Core/Billing/Premium/Models/PremiumUserSale.cs
@@ -1,6 +1,7 @@
 ﻿using Bit.Core.Billing.Models;
 using Bit.Core.Billing.Models.Sales;
 using Bit.Core.Billing.Payment.Models;
+using Bit.Core.Billing.Tax.Models;
 using Bit.Core.Entities;
 using Bit.Core.Enums;
 using Bit.Core.Models.Business;
