+++ src/Core/Billing/Premium/Services/IPremiumUserBillingService.cs
@@ -1,6 +1,7 @@
 ﻿using Bit.Core.Billing.Models;
 using Bit.Core.Billing.Payment.Models;
 using Bit.Core.Billing.Premium.Models;
+using Bit.Core.Billing.Tax.Models;
 using Bit.Core.Entities;
 
 namespace Bit.Core.Billing.Premium.Services;
