+++ src/Core/Billing/Premium/Services/PremiumUserBillingService.cs
@@ -5,7 +5,8 @@ using Bit.Core.Billing.Payment.Models;
 using Bit.Core.Billing.Payment.Services;
 using Bit.Core.Billing.Premium.Models;
 using Bit.Core.Billing.Services;
-using Bit.Core.Billing.Services.Implementations.AutomaticTax;
+using Bit.Core.Billing.Tax.Models;
+using Bit.Core.Billing.Tax.Services;
 using Bit.Core.Entities;
 using Bit.Core.Enums;
 using Bit.Core.Exceptions;
