+++ src/Core/Billing/Tax/Services/TaxService.cs
@@ -1,7 +1,7 @@
 ﻿using System.Text.RegularExpressions;
-using Bit.Core.Billing.Models;
+using Bit.Core.Billing.Tax.Models;
 
-namespace Bit.Core.Billing.Services;
+namespace Bit.Core.Billing.Tax.Services;
 
 public class TaxService : ITaxService
 {
