+++ src/Core/Services/IPaymentService.cs
@@ -1,10 +1,10 @@
 ﻿using Bit.Core.AdminConsole.Entities;
 using Bit.Core.AdminConsole.Models.Business;
 using Bit.Core.Billing.Models;
-using Bit.Core.Billing.Models.Api.Responses;
 using Bit.Core.Billing.Organizations.Requests;
 using Bit.Core.Billing.Payment.Models;
 using Bit.Core.Billing.Premium.Requests;
+using Bit.Core.Billing.Tax.Responses;
 using Bit.Core.Entities;
 using Bit.Core.Enums;
 using Bit.Core.Models.Business;
