+++ src/Core/Services/Implementations/StripePaymentService.cs
@@ -3,15 +3,15 @@ using Bit.Core.AdminConsole.Models.Business;
 using Bit.Core.Billing.Constants;
 using Bit.Core.Billing.Extensions;
 using Bit.Core.Billing.Models;
-using Bit.Core.Billing.Models.Api.Responses;
 using Bit.Core.Billing.Organizations.Models;
 using Bit.Core.Billing.Organizations.Requests;
 using Bit.Core.Billing.Payment.Models;
 using Bit.Core.Billing.Premium.Requests;
 using Bit.Core.Billing.Pricing;
 using Bit.Core.Billing.Services;
-using Bit.Core.Billing.Services.Contracts;
-using Bit.Core.Billing.Services.Implementations.AutomaticTax;
+using Bit.Core.Billing.Tax.Models;
+using Bit.Core.Billing.Tax.Responses;
+using Bit.Core.Billing.Tax.Services;
 using Bit.Core.Entities;
 using Bit.Core.Enums;
 using Bit.Core.Exceptions;
