+++ test/Core.Test/Services/StripePaymentServiceTests.cs
@@ -1,10 +1,11 @@
 ﻿using Bit.Core.Billing.Enums;
-using Bit.Core.Billing.Models.Api.Requests;
 using Bit.Core.Billing.Models.StaticStore.Plans;
 using Bit.Core.Billing.Organizations.Requests;
 using Bit.Core.Billing.Pricing;
 using Bit.Core.Billing.Services;
-using Bit.Core.Billing.Services.Contracts;
+using Bit.Core.Billing.Tax.Models;
+using Bit.Core.Billing.Tax.Requests;
+using Bit.Core.Billing.Tax.Services;
 using Bit.Core.Enums;
 using Bit.Core.Services;
 using Bit.Core.Test.Billing.Stubs;
